requirements.txt
absl-py==1.2.0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.3
cycler==0.11.0
decorator==4.4.2
dlib==19.24.0
entrypoints==0.3
facenet-pytorch==2.5.2
Flask==2.0.3
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
idna==3.4
imageio==2.15.0
imageio-ffmpeg==0.4.8
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
jedi==0.17.0
Jinja2==3.0.3
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.2.1
moviepy==1.0.3
natsort==8.2.0
networkx==2.5.1
numpy==1.19.5
oauthlib==3.2.1
opencv-python==4.2.0.34
pandas==1.1.5
parso==0.8.3
Pillow==8.4.0
proglog==0.1.10
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.1.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scipy==1.5.4
six==1.16.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.8.1
tifffile==2020.9.3
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
typing-extensions==4.1.1
urllib3==1.26.12
uWSGI==2.0.21
Werkzeug==2.0.3
yuvio==1.3.0
zipp==3.6.0
